move promise code to promise.ts
