+++ src/HighTable.tsx
@@ -3,7 +3,8 @@ import { DataFrame, PartialRow } from './dataframe.js'
 import { useInputState } from './hooks.js'
 import { Selection, areAllSelected, extendFromAnchor, isSelected, toggleAll, toggleIndex } from './selection.js'
 import TableHeader, { OrderBy, cellStyle } from './TableHeader.js'
-export { AsyncRow, DataFrame, ResolvablePromise, Row, arrayDataFrame, asyncRows, awaitRow, awaitRows, resolvablePromise, resolvableRow, sortableDataFrame, wrapPromise } from './dataframe.js'
+export { AsyncRow, DataFrame, Row, arrayDataFrame, asyncRows, awaitRow, awaitRows, resolvableRow, sortableDataFrame } from './dataframe.js'
+export { ResolvablePromise, resolvablePromise, wrapPromise } from './promise.js'
 export { rowCache } from './rowCache.js'
 export { Selection } from './selection.js'
 export { OrderBy } from './TableHeader.js'
