+++ test/dataframe.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest'
-import {
-  AsyncRow, DataFrame, Row, arrayDataFrame, awaitRows, resolvablePromise, sortableDataFrame, wrapPromise,
-} from '../src/dataframe.js'
+import { AsyncRow, DataFrame, Row, arrayDataFrame, awaitRows, sortableDataFrame } from '../src/dataframe.js'
+import { wrapPromise } from '../src/promise.js'
 
 export function wrapObject({ index, cells }: Row): AsyncRow {
   return {
@@ -11,19 +10,6 @@ export function wrapObject({ index, cells }: Row): AsyncRow {
     ),
   }
 }
-
-describe('resolvablePromise', () => {
-  it('should resolve with a value', async () => {
-    const promise = resolvablePromise<number>()
-    promise.resolve(42)
-    expect(await promise).toBe(42)
-  })
-  it('should reject with an error', async () => {
-    const promise = resolvablePromise<number>()
-    promise.reject(new Error('Failed'))
-    await expect(promise).rejects.toThrow('Failed')
-  })
-})
 
 describe('awaitRows', () => {
   it('should resolve with a row', async () => {
