+++ test/rowCache.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it, vi } from 'vitest'
-import { AsyncRow, awaitRows, wrapPromise } from '../src/dataframe.js'
+import { AsyncRow, awaitRows } from '../src/dataframe.js'
+import { wrapPromise } from '../src/promise.js'
 import { rowCache } from '../src/rowCache.js'
 
 // Mock DataFrame
